Refactor mini app into modular components
