+++ dynamic-capital-ton/apps/miniapp/lib/sections.ts
@@ -1,0 +1,7 @@
+export type SectionId =
+  | "overview"
+  | "plans"
+  | "intel"
+  | "activity"
+  | "appearance"
+  | "support";
